move gnss resources / pride utils
