+++ src/es_sfgtools/utils/__init__.py
@@ -1,2 +1,1 @@
 from .gage_data import list_survey_files
-from .gnss_resources import RemoteResource, WuhanIGS, CLSIGS, CDDIS, GSSC
